Cache *_SOVERSION and *_VERSION library version strings
My thanks to Nicolas Boulenguez and Rafael Laboissière for suggesting
this change which allows packagers to override the *_SOVERSION and
*_VERSION version strings for any plplot library in a convenient way
on the cmake command line whenever necessary for packaging (e.g., when
there is an ABI change for a particular language).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Stable = Buster) by configuring PLplot and checking the cached
results of that configuration which are stored in CMakeCache.txt.