Fixed memory leak and potential buffer overrrun in plcpstrm
When copying the buffer in plcpstrm, the buffer had new memory
allocated without freeing the old memory causing a memory leak.
In addition the buffer was accidentally allocated its old size,
not its new size, which would cause a buffer overrun if the new
buffer data was larger than the old.