Build system: Fix implementation of the test_all_qt target
The previous configuration of the test_all_qt target was implemented
with two different methods (an automated method and a non-automated
method that required maintenance whenever other relevant parts of the
build system changed).  Neither of those methods worked because of
various misspellings.  Also, the automated method had incomplete
coverage of the qt-related test targets.

For this commit I threw out the non-automated method and completely
fixed the automated method by dealing with the above spelling and
incompleteness issues.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Buster = Stable).  The actual test of this commit was
performed as part of the test for the last commit (commit b6023bf46).
See that commit message for the details.