Removed include of csa/nan.h form plgridd.c
This header was included simply to get access to a definition of
NaN, which it got by directly including ymath.h.

The header ymath.h is poorly documented (I can barely find reference
to it online) and the copies of it I can find (including on my own
Windows 10 Visual Studio system) indicate it is an internal header.
Including ymath.h on my system causes compiler errors and these
errors have been reported on the mailing list
https://sourceforge.net/p/plplot/discussion/8606/thread/318f366398/

Instead I have just switched to using the C99 NAN macro. I'm not
sure if PLPlot ever required supporting C89, but other C99 macros
suche as HUGE_VAL are used elsewhere in the library, so I'm sure
this is fine.